backend/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
geographiclib==1.52
geopy==2.2.0
haversine==2.5.1
numpy==1.22.3
pymongo==4.1.1
python-slugify==6.1.2
pytz==2022.1
secure-smtplib==0.1.1
sqlparse==0.2.4
text-unidecode==1.3
tzdata==2022.1
